refactor: Simpler and more type-safe `BazelTaskInfo`
* Add the `bazelTaskInfo` member to the `BazelTaskDefinition` type so it
  can be accessed in a more typesafe manner.
* Merges `bazel_task_info.ts` into `tasks.ts`. Otherwise, adding the
  `bazelTaskInfo` member would have introduced a cyclical dependency.
  Also, the `bazel_task_info.ts` was rather small anyway, so keeping
  this in a separate file didn't provide much value anyway.

Refactoring in preparation for #362
